style: 캘린더 데일리룩 에디터 카테고리별 Default 위치 조정함수 분리
